@@ -11,7 +11,7 @@ plugins {
 
 subprojects {
     group "com.github.ixtf.japp"
-    version "1.0.0"
+    version "3.0.0"
 
     apply plugin: "java-library"
     apply plugin: "maven-publish"
@@ -44,8 +44,8 @@ subprojects {
     }
 
     java {
-        sourceCompatibility = JavaVersion.VERSION_11
-        targetCompatibility = JavaVersion.VERSION_11
+        sourceCompatibility = JavaVersion.VERSION_17
+        targetCompatibility = JavaVersion.VERSION_17
         withJavadocJar()
         withSourcesJar()
     }
