@@ -5,12 +5,13 @@ import com.github.ixtf.api.Util;
 import com.google.common.collect.ImmutableMap;
 import com.google.inject.Inject;
 import com.google.inject.name.Named;
-import graphql.ExecutionResult;
 import graphql.GraphQL;
 import io.opentracing.Span;
 import io.opentracing.Tracer;
 import io.opentracing.tag.Tags;
-import io.vertx.core.*;
+import io.vertx.core.AbstractVerticle;
+import io.vertx.core.Handler;
+import io.vertx.core.Promise;
 import io.vertx.core.buffer.Buffer;
 import io.vertx.core.eventbus.Message;
 import io.vertx.core.json.JsonArray;
@@ -19,16 +20,12 @@ import io.vertx.ext.web.handler.graphql.impl.GraphQLBatch;
 import io.vertx.ext.web.handler.graphql.impl.GraphQLInput;
 import io.vertx.ext.web.handler.graphql.impl.GraphQLQuery;
 import lombok.extern.slf4j.Slf4j;
-import reactor.core.publisher.Mono;
 
 import java.util.Optional;
-import java.util.stream.StreamSupport;
 
 import static com.github.ixtf.api.guice.ApiModule.GRAPHQL_ADDRESS;
 import static com.github.ixtf.guice.GuiceModule.injectMembers;
 import static java.util.Optional.ofNullable;
-import static java.util.stream.Collectors.collectingAndThen;
-import static java.util.stream.Collectors.toList;
 
 @Slf4j
 public class GraphqlVerticle extends AbstractVerticle implements Handler<Message<Buffer>> {
@@ -56,41 +53,41 @@ public class GraphqlVerticle extends AbstractVerticle implements Handler<Message
     @Override
     public void handle(Message<Buffer> reply) {
         final var spanOpt = spanOpt(reply);
-        Mono.fromCompletionStage(() -> {
+        try {
             final var graphQLInput = GraphQLInput.decode(reply.body());
-            if (graphQLInput instanceof GraphQLBatch) {
-                return handleBatch((GraphQLBatch) graphQLInput).map(JsonArray::toBuffer).toCompletionStage();
-            } else if (graphQLInput instanceof GraphQLQuery) {
-                return handleQuery((GraphQLQuery) graphQLInput).map(JsonObject::toBuffer).toCompletionStage();
+            if (graphQLInput instanceof final GraphQLQuery query) {
+                final var jsonObject = handleQuery(reply, query);
+                reply.reply(jsonObject.toBuffer());
+            } else if (graphQLInput instanceof final GraphQLBatch batch) {
+                final var jsonArray = handleBatch(reply, batch);
+                reply.reply(jsonArray.toBuffer());
+            } else {
+                reply.fail(400, "no GraphQLInput");
             }
-            throw new RuntimeException();
-        }).subscribe(it -> {
-            reply.reply(it);
-            spanOpt.ifPresent(Span::finish);
-        }, e -> {
-            reply.fail(400, e.getMessage());
+        } catch (Throwable e) {
+            final var errorMessage = e.getMessage();
+            reply.fail(400, errorMessage);
             log.error("", e);
-            spanOpt.ifPresent(span -> span.setTag(Tags.ERROR, true).log(e.getMessage()).finish());
-        });
+            spanOpt.ifPresent(span -> span.setTag(Tags.ERROR, true).log(errorMessage).finish());
+        } finally {
+            spanOpt.ifPresent(Span::finish);
+        }
     }
 
-    private Future<JsonObject> handleQuery(GraphQLQuery query) {
-        final var completableFuture = graphQL.executeAsync(builder -> {
+    private JsonObject handleQuery(Message<Buffer> reply, GraphQLQuery query) {
+        final var executionResult = graphQL.execute(builder -> {
             builder.query(query.getQuery());
             ofNullable(query.getOperationName()).filter(J::nonBlank).ifPresent(builder::operationName);
             ofNullable(query.getVariables()).ifPresent(builder::variables);
+            builder.graphQLContext(ctx -> reply.headers().forEach(entry -> ctx.of(entry.getKey(), entry.getValue())));
             return builder;
         });
-        return Future.fromCompletionStage(completableFuture, vertx.getOrCreateContext())
-                .map(ExecutionResult::toSpecification)
-                .map(JsonObject::new);
+        return new JsonObject(executionResult.toSpecification());
     }
 
-    private Future<JsonArray> handleBatch(GraphQLBatch batch) {
-        return StreamSupport.stream(batch.spliterator(), false)
-                .map(q -> (Future) handleQuery(q))
-                .collect(collectingAndThen(toList(), CompositeFuture::all))
-                .map(CompositeFuture::list)
-                .map(JsonArray::new);
+    private JsonArray handleBatch(Message<Buffer> reply, GraphQLBatch batch) {
+        final var jsonArray = new JsonArray();
+        batch.forEach(it -> jsonArray.add(handleQuery(reply, it)));
+        return jsonArray;
     }
 }
